Adds example with category prop
As was noted in #9, there are some extra things that could be added to
the data structure, category being one of them.

Thought I would test out a design where the category is just above
the definition and a smidge smaller.

One thing to note is the redundant list role on the `<ul>`. Added this because the
rule `list-style-type: none;` can mess with some screen readers